Add files via upload
Visualization generated using Bokeh. 

The text on LinkedIn (see 3b) reads:

"Here is an automatically generated visualization of Rewa centered around the Rani Talab Kalika Mandir (the generated image has been cropped appropriately, post-production, as it were). The image was pulled out of Google Maps, and glyphs laid over on top. We can easily change the type of the map with a simple change in the parameter file as well."